requirements.txt
streamlit
pandas
sqlalchemy
python-docx
pdfplumber
spacy
scikit-learn
